chore: dropping useless `PrivateKernelOracle` interface
We had `PrivateKernelOracle` interface and then we had it implemented only by one class that was called `PrivateKernelOracleImpl`. That's just bizarre useless disgusting boilerplate that cannot be tolerated so I dropped it.
