+++ yarn-project/pxe/src/private_kernel/private_kernel_execution_prover.test.ts
@@ -158,9 +158,14 @@ describe('Private Kernel Sequencer', () => {
     oracle.getVkMembershipWitness.mockResolvedValue(MembershipWitness.random(VK_TREE_HEIGHT));
 
     oracle.getContractAddressPreimage.mockResolvedValue({
+      version: 1 as const,
+      salt: Fr.random(),
+      deployer: await AztecAddress.random(),
       currentContractClassId: Fr.random(),
       originalContractClassId: Fr.random(),
+      initializationHash: Fr.random(),
       publicKeys: await PublicKeys.random(),
+      address: await AztecAddress.random(),
       saltedInitializationHash: Fr.random(),
     });
     oracle.getContractClassIdPreimage.mockResolvedValue({
